Fix: Move DBQueries class
